feat: Support wgrad in MoE overlap (#278)
We implemented a separate and different version of Megatron's
delay_wgrad_compute. This version does not use the Megatron modules'
backward_dw() and works around the following difficulties:

(1) backward_dw() has a hard-coded TE version requirement (>3.0); it is
unavailable in our existing docker releases.
(2) Megatron's delay_wgrad_compute requires support for various ops such
as layer_norm, gemm, attn, grouped gemm etc, which are not available in
our TE or in primus-turbo. Particularly, attn.backward_dw() is used by
Megatron to overlap with PP comm, but its support in Primus Turbo would
be challenging to implement in the underlying C++ backends;

Instead, in our implementation we use grouped_gemm.backward_dw() to
overlap with PP comm.

We observe it is able to well overlap with PP comm on top of existing
MoE overlaps. It gives us additional 5% performance gain in our
prototype testing.

To use this feature, set the following parameters:
      overlap_moe_expert_parallel_comm: true
      patch_moe_overlap: true
      delay_wgrad_compute: false